+++ packages/app/src/components/VisualizationOptions/Options/CompletedOnly.js
@@ -1,16 +1,20 @@
 import React from 'react'
 
 import i18n from '@dhis2/d2-i18n'
+import { Label } from '@dhis2/ui-core'
 
 import CheckboxBaseOption from './CheckboxBaseOption'
 
 const CompletedOnly = () => (
-    <CheckboxBaseOption
-        label={i18n.t('Include only completed events')}
-        option={{
-            name: 'completedOnly',
-        }}
-    />
+    <div>
+        <Label>{i18n.t('Event data')}</Label>
+        <CheckboxBaseOption
+            label={i18n.t('Only include completed events')}
+            option={{
+                name: 'completedOnly',
+            }}
+        />
+    </div>
 )
 
 export default CompletedOnly