+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [34.0.4](https://github.com/dhis2/data-visualizer-app/compare/v34.0.3...v34.0.4) (2020-02-14)
+
+
+### Bug Fixes
+
+* owner filter label (DHIS2-8263) ([#659](https://github.com/dhis2/data-visualizer-app/issues/659)) ([591bbe5](https://github.com/dhis2/data-visualizer-app/commit/591bbe5474a118db27a9b45174dbc87a692f2df8))
+
+## [34.0.3](https://github.com/dhis2/data-visualizer-app/compare/v34.0.2...v34.0.3) (2020-02-14)
+
+
+### Bug Fixes
+
+* data with multiple indicators error message ([#658](https://github.com/dhis2/data-visualizer-app/issues/658)) ([f200787](https://github.com/dhis2/data-visualizer-app/commit/f2007875125986c29c908c28b6b50c6f84c72852))
+
+## [34.0.2](https://github.com/dhis2/data-visualizer-app/compare/v34.0.1...v34.0.2) (2020-02-13)
+
+
+### Bug Fixes
+
+* missing Pivot options ([#665](https://github.com/dhis2/data-visualizer-app/issues/665)) ([6aea633](https://github.com/dhis2/data-visualizer-app/commit/6aea6330ae370697a4c30418a50777d14c31e40f))
+
 ## [34.0.1](https://github.com/dhis2/data-visualizer-app/compare/v34.0.0...v34.0.1) (2020-02-13)